map loop structure

+++ src/GamePlayer/GameView/src/Player/GamePlayVisualization.java
@@ -15,9 +15,9 @@ import javafx.util.Duration;
 
 public class GamePlayVisualization extends Application {
     private String Title = "VoogaSalad Game";
-    private static final int FRAMES_PER_SECOND = 1;
-    private static final int MILLISECOND_DELAY = 1000 / FRAMES_PER_SECOND;
-    public static final double SECOND_DELAY = 1.0 / FRAMES_PER_SECOND;
+    private static final long FRAMES_PER_SECOND = 1;
+    private static final long MILLISECOND_DELAY = 1000 / FRAMES_PER_SECOND;
+    public static final long SECOND_DELAY = 1 / FRAMES_PER_SECOND;
     private static final Paint backgroundColor = Color.NAVY;
     private double screenWidth;
     private double screenHeight;
